+++ packages/core/src/utils/breakpoints.ts
@@ -82,7 +82,10 @@ const builtInStylesWithDesignTokens = [
   'cfWidth',
   'cfHeight',
   'cfBackgroundColor',
+  'cfBorder',
   'cfFontSize',
+  'cfLineHeight',
+  'cfLetterSpacing',
 ];
 
 export const getValueForBreakpoint = (
@@ -101,7 +104,7 @@ export const getValueForBreakpoint = (
         const breakpointValue =
           valuesByBreakpoint[breakpointId] || valuesByBreakpoint[fallbackBreakpointId];
 
-        return getDesignTokenRegistration(breakpointValue);
+        return getDesignTokenRegistration(breakpointValue, variableName);
       }
       if (valuesByBreakpoint[breakpointId]) {
         // If the value is defined, we use it and stop the breakpoints cascade