+++ packages/experience-builder-sdk/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [4.0.0-alpha.7](https://github.com/contentful/experience-builder/compare/@contentful/experience-builder@4.0.0-alpha.6...@contentful/experience-builder@4.0.0-alpha.7) (2024-01-10)
+
+**Note:** Version bump only for package @contentful/experience-builder
+
+# [4.0.0-alpha.6](https://github.com/contentful/experience-builder/compare/@contentful/experience-builder@4.0.0-alpha.5...@contentful/experience-builder@4.0.0-alpha.6) (2024-01-10)
+
+**Note:** Version bump only for package @contentful/experience-builder
+
+# [4.0.0-alpha.5](https://github.com/contentful/experience-builder/compare/@contentful/experience-builder@4.0.0-alpha.4...@contentful/experience-builder@4.0.0-alpha.5) (2024-01-09)
+
+**Note:** Version bump only for package @contentful/experience-builder
+
+# [4.0.0-alpha.4](https://github.com/contentful/experience-builder/compare/@contentful/experience-builder@4.0.0-alpha.3...@contentful/experience-builder@4.0.0-alpha.4) (2024-01-09)
+
+**Note:** Version bump only for package @contentful/experience-builder
+
 # [4.0.0-alpha.3](https://github.com/contentful/experience-builder/compare/@contentful/experience-builder@4.0.0-alpha.2...@contentful/experience-builder@4.0.0-alpha.3) (2024-01-09)
 
 ### Bug Fixes