+++ packages/storybook-addon/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.2-alpha.8](https://github.com/contentful/experience-builder-storybook/compare/@contentful/experience-builder-storybook-addon@0.0.2-alpha.7...@contentful/experience-builder-storybook-addon@0.0.2-alpha.8) (2024-01-10)
+
+**Note:** Version bump only for package @contentful/experience-builder-storybook-addon
+
+## [0.0.2-alpha.7](https://github.com/contentful/experience-builder-storybook/compare/@contentful/experience-builder-storybook-addon@0.0.2-alpha.6...@contentful/experience-builder-storybook-addon@0.0.2-alpha.7) (2024-01-10)
+
+**Note:** Version bump only for package @contentful/experience-builder-storybook-addon
+
+## [0.0.2-alpha.6](https://github.com/contentful/experience-builder-storybook/compare/@contentful/experience-builder-storybook-addon@0.0.2-alpha.5...@contentful/experience-builder-storybook-addon@0.0.2-alpha.6) (2024-01-09)
+
+**Note:** Version bump only for package @contentful/experience-builder-storybook-addon
+
+## [0.0.2-alpha.5](https://github.com/contentful/experience-builder-storybook/compare/@contentful/experience-builder-storybook-addon@0.0.2-alpha.4...@contentful/experience-builder-storybook-addon@0.0.2-alpha.5) (2024-01-09)
+
+**Note:** Version bump only for package @contentful/experience-builder-storybook-addon
+
 ## [0.0.2-alpha.4](https://github.com/contentful/experience-builder-storybook/compare/@contentful/experience-builder-storybook-addon@0.0.2-alpha.3...@contentful/experience-builder-storybook-addon@0.0.2-alpha.4) (2024-01-09)
 
 **Note:** Version bump only for package @contentful/experience-builder-storybook-addon