+++ packages/visual-editor/CHANGELOG.md
@@ -3,6 +3,26 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.2-alpha.8](https://github.com/contentful/experience-builder/compare/@contentful/experience-builder-visual-editor@0.0.2-alpha.7...@contentful/experience-builder-visual-editor@0.0.2-alpha.8) (2024-01-10)
+
+**Note:** Version bump only for package @contentful/experience-builder-visual-editor
+
+## [0.0.2-alpha.7](https://github.com/contentful/experience-builder/compare/@contentful/experience-builder-visual-editor@0.0.2-alpha.6...@contentful/experience-builder-visual-editor@0.0.2-alpha.7) (2024-01-10)
+
+### Bug Fixes
+
+- handle missing linkType property ([#237](https://github.com/contentful/experience-builder/issues/237)) ([1827d96](https://github.com/contentful/experience-builder/commit/1827d960f0a53412e1e39b06e265f7e1511e54f5))
+
+## [0.0.2-alpha.6](https://github.com/contentful/experience-builder/compare/@contentful/experience-builder-visual-editor@0.0.2-alpha.5...@contentful/experience-builder-visual-editor@0.0.2-alpha.6) (2024-01-09)
+
+### Bug Fixes
+
+- **visual-editor:** remove error around passing editor props to normal elements ([#231](https://github.com/contentful/experience-builder/issues/231)) ([4024a68](https://github.com/contentful/experience-builder/commit/4024a68b3da06919ae777c15e7d25a8a40d6c263))
+
+## [0.0.2-alpha.5](https://github.com/contentful/experience-builder/compare/@contentful/experience-builder-visual-editor@0.0.2-alpha.4...@contentful/experience-builder-visual-editor@0.0.2-alpha.5) (2024-01-09)
+
+**Note:** Version bump only for package @contentful/experience-builder-visual-editor
+
 ## [0.0.2-alpha.4](https://github.com/contentful/experience-builder/compare/@contentful/experience-builder-visual-editor@0.0.2-alpha.3...@contentful/experience-builder-visual-editor@0.0.2-alpha.4) (2024-01-09)
 
 ### Bug Fixes